fix tiny gcc warning
git-svn-id: http://vtp.googlecode.com/svn/trunk@5838 2ba57974-ef2e-0410-970c-f3cb55f6a86a

+++ TerrainSDK/vtui/LinearStructDlg.cpp
@@ -459,7 +459,7 @@ void LinearStructureDlg::SlidersToValues(int which)
 		m_param.m_fConnectWidth  = WIDTH_MIN + m_iConnWidth *   (WIDTH_MAX - WIDTH_MIN) / 100.0f;
 		break;
 	case ID_SLOPE_SLIDER:
-		m_param.m_iConnectSlope  = SLOPE_MIN + m_iSlope *   (SLOPE_MAX - SLOPE_MIN) / 100.0f;
+		m_param.m_iConnectSlope  = SLOPE_MIN + (int) (m_iSlope * (SLOPE_MAX - SLOPE_MIN) / 100.0f);
 		break;
 	}
 }
